// server/types.ts
export interface User {
  id: number
  uuid: string
  email: string
  password: string
  avatar?: string
  first_name: string
  last_name: string
  email_verified?: boolean
  is_active?: boolean
  last_login?: Date
  phone?: string
  created_at: Date
  deleted_at?: Date
  updated_at?: Date
  role_id: number
  metadata?: JSON
  notifications?: JSON
  status?: UserStatus
  role: Role
  twofactor?: TwoFactor
  author?: Author
  chats: Chat[]
  notification: Notification[]
  posts: Post[]
  comments: Comment[]
}

export interface CustomWebSocket {
  send: (
    message: string | ArrayBuffer,
    isBinary?: boolean,
    compress?: boolean,
  ) => void
  end: (data: any) => void
}

export type UserStatus = 'ACTIVE' | 'INACTIVE' | 'BANNED' | 'SUSPENDED'
// ...
export interface Role {
  id: number
  name: string
  permissions: Permission[]
  rolePermissions: RolePermission[]
}

export interface Permission {
  id: number
  name: string
  rolePermissions: RolePermission[]
}

export interface RolePermission {
  id: number
  role: Role
  permission: Permission
}
// ...
export interface Notification {
  id: number
  user_id: number
  type: string | null
  title: string
  message: string
  link: string | null
  created_at: Date | null
  updated_at: Date | null
}

export interface Wallet {
  id: number
  uuid: string
  type: WalletType
  user_id: number
  balance: number
  inOrder: number
  currency: string
  addresses: any
  created_at: string
  updated_at: string
  transactions: Transaction[]
  invoices: Invoice[]
}

export enum WalletType {
  FIAT = 'FIAT',
  SPOT = 'SPOT',
  ECO = 'ECO',
}
// ...
export enum TransactionType {
  FAILED = 'FAILED',
  DEPOSIT = 'DEPOSIT',
  WITHDRAW = 'WITHDRAW',
  OUTGOING_TRANSFER = 'OUTGOING_TRANSFER',
  INCOMING_TRANSFER = 'INCOMING_TRANSFER',
  PAYMENT = 'PAYMENT',
  REFUND = 'REFUND',
  BINARY_ORDER = 'BINARY_ORDER',
  EXCHANGE_ORDER = 'EXCHANGE_ORDER',
  INVESTMENT = 'INVESTMENT',
  INVESTMENT_ROI = 'INVESTMENT_ROI',
  AI_INVESTMENT = 'AI_INVESTMENT',
  AI_INVESTMENT_ROI = 'AI_INVESTMENT_ROI',
  INVOICE = 'INVOICE',
  FOREX_DEPOSIT = 'FOREX_DEPOSIT',
  FOREX_WITHDRAW = 'FOREX_WITHDRAW',
  FOREX_INVESTMENT = 'FOREX_INVESTMENT',
  FOREX_INVESTMENT_ROI = 'FOREX_INVESTMENT_ROI',
  ICO_CONTRIBUTION = 'ICO_CONTRIBUTION',
  REFERRAL_REWARD = 'REFERRAL_REWARD',
  STAKING = 'STAKING',
  STAKING_REWARD = 'STAKING_REWARD',
  P2P_OFFER_TRANSFER = 'P2P_OFFER_TRANSFER',
  P2P_TRADE = 'P2P_TRADE',
}

export enum TransactionStatus {
  PENDING = 'PENDING',
  COMPLETED = 'COMPLETED',
  FAILED = 'FAILED',
  CANCELLED = 'CANCELLED',
  EXPIRED = 'EXPIRED',
  REJECTED = 'REJECTED',
  REFUNDED = 'REFUNDED',
  TIMEOUT = 'TIMEOUT',
}

// The Transaction type maps to your Prisma model for transactions
export interface Transaction {
  id: number
  uuid: string
  user_id: number
  wallet_id: number
  type: TransactionType
  status: TransactionStatus
  amount: number
  fee?: number
  description?: string
  metadata?: any // Replace with a more specific type if you know the shape of the metadata
  reference_id?: string
  created_at: Date
  updated_at?: Date
  wallet: Wallet
  user: User
  invoice: Invoice
}

export interface Invoice {
  id: number
  wallet_id: number
  user_id: number
  amount: number
  status: string
  created_at: string
  updated_at: string
}

export interface TwoFactor {
  id: number
  user_id: number
  secret: string
  type: TwoFactorType
  created_at: string
  updated_at: string
}

export type TwoFactorType = 'EMAIL' | 'SMS' | 'APP'

export interface Author {
  id: number
  uuid: string
  user_id: number
  status: AuthorStatus
  user: User
  posts: Post[]
}

export type AuthorStatus = 'PENDING' | 'APPROVED' | 'REJECTED'

export interface Category {
  id: number
  name: string
  slug: string
  created_at: Date | null
  updated_at: Date | null
  post: Post[]
}

export interface Comment {
  id: number
  content: string
  author_id: number
  author: Author
  post_id: number
  post: Post
  created_at: Date | null
  updated_at: Date | null
}

export interface PostTag {
  id: number
  post_id: number
  tag_id: number
  created_at: Date | null
  updated_at: Date | null
}

export interface Post {
  id: number
  title: string
  content: string
  category_id: number
  author_id: number
  author: Author
  created_at: Date | null
  updated_at: Date | null
  comment: Comment[]
  post_tag: PostTag[]
  status: PostStatus
}

export type PostStatus = 'PUBLISHED' | 'DRAFT' | 'TRASH'
// ...
export interface Chat {
  id: number
  uuid: string
  user_id: number
  user: User
  agent_id: number
  agent: User
  messages: Message[] | null
}
// ...
export interface Message {
  type: string
  text: string
  time: string
  user_id: number
  agent_id: number
  attachments: Attachment[]
}
// ...
export enum AiTradingResult {
  WIN = 'WIN',
  LOSS = 'LOSS',
  DRAW = 'DRAW',
}

export enum AiTradingStatus {
  PENDING = 'PENDING',
  ACTIVE = 'ACTIVE',
  COMPLETED = 'COMPLETED',
  CANCELLED = 'CANCELLED',
  REJECTED = 'REJECTED',
}

export enum AiTradingTimeframe {
  HOUR = 'HOUR',
  DAY = 'DAY',
  WEEK = 'WEEK',
  MONTH = 'MONTH',
}

export interface AiTradingPlan {
  id: number
  name: string
  title: string
  description?: string
  image?: string
  status?: boolean
  invested: number
  profit_percentage: number
  min_profit: number
  max_profit: number
  min_amount: number
  max_amount: number
  trending?: boolean
  default_profit: number
  default_result: AiTradingResult
}

export interface AiTrading {
  id: number
  uuid: string
  user_id: number
  user: User
  plan_id: number
  plan: AiTradingPlan
  duration_id: number
  duration: AiTradingDuration
  market: string
  amount: number
  profit?: number
  result?: AiTradingResult
  status: AiTradingStatus
  created_at: Date
  updated_at: Date
}

export interface AiTradingDuration {
  id: number
  duration: number
  timeframe: AiTradingTimeframe
  ai_trading_plan_duration: AiTradingPlanDuration[]
}

export interface AiTradingPlanDuration {
  id: number
  plan_id: number
  plan: AiTradingPlan
  duration_id: number
  duration: AiTradingDuration
}